+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - State.matches(observable, predicate) added.
 ### is.codion.common.model
 - FilterModel.Items.sort() added.
+- DefaultMultiSelection and DefaultSingleSelection added along with MultiSelection.IndexedItems added.
 ### is.codion.framework.db.http
 - HttpTransport and JdkHttpTransport added, service based
 ### is.codion.swing.framework.ui